*** empty log message ***
git-svn-id: file:///usr/local/svn/fsc2@3532 934aee4a-8510-0410-ae50-ec5b2ea4bb86

+++ trunk/fsc2/src/func.h
@@ -58,7 +58,7 @@ typedef struct
 {
 	const char *name;                 /* name of the function */
     Var * ( * fnct )( Var * );        /* pointer to the function */
-	long nargs;                       /* number of arguments */
+	int nargs;                        /* number of arguments */
     int access_flag;                  /* asscessibility flag */
 	Device *device;                   /* handle of defining device module */
 	bool to_be_loaded;                /* set if function has to be loaded */
