+++ trunk/fsc2/src/variables.h
@@ -71,8 +71,9 @@ typedef struct Var_
 	} val;
 	int    dim;              /* dimension of array */
 	int    *sizes;           /* array of sizes of dimensions */
-	long   len;              /* total len of array */
+	size_t   len;            /* total len of array */
 	long   flags;
+
 	struct Var_ *from;
 	struct Var_ *next;       /* next variable in list or stack */
 	struct Var_ *prev;       /* previous variable in list or stack */
